LayoutTests/platform/win-wk2/http/tests/inspector-enabled/database-open-expected.txt

CONSOLE MESSAGE: line 12: SECURITY_ERR: DOM Exception 18: An attempt was made to break through the security policy of the user agent.
FAIL: Timed out waiting for notifyDone to be called
